Avoid suggesting equivalencies we want to avoid
Since we do not want the reader to consider fetch and pull as
equivalencies, it is better to avoid putting the idea in their head at
all. Plus, the expanded introduction now makes it clear that pull is
doing two things in a single command, so the comment is now redundant
and we can simplify the text by removing the sentence.