Define tracking branches
THe reader may be new to git and unsure about what a tracking branch is
and how (if at all) it differs from a local branch. Adding a note at the
beginning will help address that potential confusion. The note also
makes it clear that the commands only works for tracked branches so we
can assume any local branches discussed later in the text are tracking
branches without having to be explicit about that fact every time we
talk about the local branch.

The note style used is standard GitHub markdown for notes, which renders
a small information icon when rendered in the browser.